First reworked version - implementations in separate files for Windows, UNIX and C++11 systems
