+++ sleep.h
@@ -1,0 +1,10 @@
+#ifndef SLEEP_H
+#define SLEEP_H
+
+#include <nan.h>
+
+void node_sleep(unsigned seconds);
+void node_usleep(unsigned usec);
+void node_msleep(unsigned msec);
+
+#endif SLEEP_H
